requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref==3.5.2
dj-database-url==1.0.0
django==4.1
django-heroku==0.3.1
gunicorn==20.1.0
pillow==9.2.0
pip==22.1.2
psycopg2==2.9.3
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==63.4.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.1
wheel==0.37.1
whitenoise==6.2.0
django-environ==0.4.5
tweepy==4.10.1
cloudinary==1.29.0
django-mdeditor==0.1.20
Markdown==3.4.1